requirements.txt
streamlit==1.24.0
langchain==0.0.228
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
instaloader==4.10.0